Port TestFixedZoneRules to scalatest
